refactor: remove unuseful css files
