--- src/components/forms/uncontrolled/Uncontrolled.module.css
@@ -1,42 +1,0 @@
-.form {
-    width: 50vw;
-    display: flex;
-    flex-direction: column;
-    gap: 40px;
-    overflow-y: scroll;
-}
-
-.form > h1 {
-  color: black;
-}
-
-.form > button {
-    padding: 15px 30px;
-    border-radius: 10px;
-    border: none;
-    background-color: var(--btn-color-red);
-    font-size: 16px;
-    color: white;
-    align-self: center;
-    cursor: pointer;
-}
-
-.form > button:hover {
-    background-color: var(--btn-color-red-hover);
-}
-
-.content {
-    gap: 100px;
-    display: flex;
-    justify-content: space-between;
-}
-
-.section {
-    display: flex;
-    flex-direction: column;
-    flex: 1;
-    flex-basis: 50%;
-    gap: 20px;
-    background-color: white;
-    color: black;
-}
