+++ src/components/forms/uncontrolled/inputs/email/EmailInput.tsx
@@ -1,4 +1,4 @@
-import styles from '../Inputs.module.css';
+import styles from '../../../Input.module.css';
 
 function EmailInput() {
   return (
